optical-flow-transfer/requirements.txt
tqdm
Pillow
python-opencv
numpy